feat(stanza): split stanza
