+++ src/lib.rs
@@ -8,7 +8,7 @@ extern crate tokio_io;
 extern crate bytes;
 
 
-use connect::{AUTH, PLAIN, ServerMessage};
+use connect::{ServerMessage};
 
 use base64::{encode};
 use futures::{Future, Stream, Sink};
@@ -26,6 +26,7 @@ use tokio_tls::TlsStream;
 
 mod connect;
 mod codec;
+mod stanza;
 
 pub use connect::ClientMessage;
 use codec::LineCodec;
@@ -57,7 +58,7 @@ pub fn connect_client<F>(out_tx: mpsc::Sender<(ClientMessage, mpsc::Sender<Clien
 
     let starttls = |(response, t): (Option<String>, Framed<TokioStream, LineCodec>)| {
         println!("Response START: {:?}", response);
-        t.send(AUTH.to_string())
+        t.send(stanza::non_stanza::AUTH.to_string())
             .and_then(|transport| transport.into_future().map_err(|(e, _)| e))
     };
 
@@ -67,7 +68,6 @@ pub fn connect_client<F>(out_tx: mpsc::Sender<(ClientMessage, mpsc::Sender<Clien
         let cx = builder.build().unwrap();
 
         println!("connected");
-        // cx.connect_no_domain(transport.into_inner()).map_err(|e| {
         cx.connect_async(domain, transport.into_inner()).map_err(|e| {
             io::Error::new(io::ErrorKind::Other, e)
         })
@@ -96,7 +96,7 @@ pub fn connect_client<F>(out_tx: mpsc::Sender<(ClientMessage, mpsc::Sender<Clien
 
             let bytes = str::from_utf8(&data).unwrap().as_bytes();
             let plain = encode(bytes);
-            let plain = format!("{}{}</auth>", PLAIN, plain);
+            let plain = format!("{}{}</auth>", stanza::non_stanza::PLAIN, plain);
             transport.send(plain)
         })
     .and_then(|transport| transport.into_future().map_err(|(e, _)| e))
