+++ src/stanza/non_stanza.rs
@@ -1,0 +1,4 @@
+use std::str;
+
+pub const AUTH: &'static str = "<starttls xmlns='urn:ietf:params:xml:ns:xmpp-tls'/>";
+pub const PLAIN: &'static str = "<auth xmlns='urn:ietf:params:xml:ns:xmpp-sasl' mechanism='PLAIN'>";
